Fixing errors in account feed event so that the object store can handle it correctly.
